Removing deprecated /clients
